refactor: split sample API routes
