+++ src/api/__init__.py
@@ -1,11 +1,14 @@
-"""API package providing FastAPI application and route definitions."""
-
-from .sample_routes import app
+"""API package providing FastAPI route definitions."""
 
 try:  # pragma: no cover - optional dependency
     from .routes import router
 except Exception:  # ImportError if FastAPI is missing
     router = None  # type: ignore
 
-__all__ = ["app", "router"]
+try:  # pragma: no cover - optional dependency
+    from .sample_routes import router as sample_router
+except Exception:  # ImportError if FastAPI is missing
+    sample_router = None  # type: ignore
+
+__all__ = ["router", "sample_router"]
