+++ tests/test_api_routes.py
@@ -1,24 +1,24 @@
 import sys
-import sys
 from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
+sys.path.insert(0, str(ROOT / "src"))
 
 from src.api.sample_routes import api_provision, api_subgraph, api_treatment
 
 
 def test_subgraph_endpoint():
-    data = api_subgraph(["doc1", "doc2"], limit=10, offset=0)
-    assert len(data["nodes"]) == 2
-    assert any(e["target"] == "doc2" for e in data["edges"])
+    data = api_subgraph("example text")
+    assert "cloud" in data
 
 
 def test_treatment_endpoint():
-    data = api_treatment(doc="doc1", limit=10, offset=0)
-    assert any(e["target"] == "doc2" for e in data["treatments"])
+    data = api_treatment("a person shall act")
+    assert "rules" in data
 
 
 def test_provision_endpoint():
-    data = api_provision(doc="doc1", id="prov1")
-    assert data["identifier"] == "prov1"
+    data = api_provision("Sample provision text")
+    assert "provision" in data
+
